chore: Moved datasource exportable to appropriate package (#31046)
